requirements.txt
flask
SQLAlchemy
psycopg2-binary
python-dotenv
